+++ src/App/App.css
@@ -1,4 +1,4 @@
-@import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
+@import '../../node_modules/bootstrap/dist/css/bootstrap.min.css';
 
 .container {  
   margin:0;
