+++ src/App/App.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import Calculator from '../Calculator/Calculator';
+import './App.css';
+
+function App() {
+  return (    
+      <Calculator />    
+  )
+}
+
+export default App;
